Pathfinding bug fix
The AI snakes mostly ended up travelling into walls. This is because they couldn't switch back on themselves and the closest food was in the same x or y position but in the opposite direction. This was rectified by choosing randomly to change to left and right if travelling up or down and vice versa.