StateManager and BaseState Rework
The StateManager and BaseState classes have been reworked to use references over pointers. This ensures that They are accessing what they are meant to and, since each thing they are going to be accessing will exist, references should have been used in the first place.

On top of this, sicne each state will be using sound of some sort, each state was given a reference to the SoundManager class