StateManager Overhaul
The state manager now works on enums as opposed to claling the state pointer immediately, following a code review with Third Year CGP student Ryan Swann, where he pointed out that the game could crash if a nullptr was the argument.

The SoundManager is now a reference in every instance I used it, opposed to the pointer that it was.

The BaseState Class was moved into its own file, for ease of use.

sfml-main libraries were added so that the game could be built to a .exe

Default constructors were made for the states.

Essentially the code is a lot better now!