Changes, Fixes and Documentation
After playtesting the game, several areas were brought up that needed tweaking and fixing.

Only 1 Gobble Mode can spawn at once
Gobble Mode ends when the snake that had it died
The timer remains paused when the game is paused
You can't turn when paused any more
Linked list was remade
Collision was redone to be more streamlined
Occasionally the game would freeze if a head on collision occured, that has been fixed.
A victory noise has been added if the player wins in single player, as well as a failure noise if the player loses.
The code has been properly documented!